Refactor AbstractObjectProvider
Previously, the `AbstractObjectProvider` only worked with
`Identifiables`, which made it incompatible with the new version of the
AAS metamodel.

This renames and restructures both the `AbstractObjectProvider` and the
`AbstractObjectStore`. In all non-abstract subclasses where `Object`
appears in the class or method name, it has been replaced with
`Identifiable`. Old classes are still available with a deprecation
warning.

Moreover, `AbstractObjectProvider` and `AbstractObjectStore` are now
generic to be able to handle more classes than just `Identifiables`.

In order to handle the new `AASDescriptor`, a new class `HasIdentifier`
has been added. It is intended to be an abstract superclass for all
classes that have an identifier, but are not `Identifiables`.

As of now, this PR is intended to serve as a basis for discussion.
Therefore, the documentation has not yet been adapted.

Fixes #428